refactor: move OPTIONS_MODES to product_model
Cause it's more related to model than to it's representation
