+++ src/components/product/Product.stories.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import { Product, OPTIONS_MODES } from './Product'
+import { Product } from './Product'
+import { OPTIONS_MODES } from '../../models/product_model'
 
 const parameters = [
   {
